requirements/common.txt
python-dateutil==2.7.3
pytz==2018.5
requests==2.19.1

# Testing
mock==1.3.0
vcrpy==1.7.4
